requirements.txt
yfinance>=0.2.40
pandas>=2.2
numpy>=1.26
statsmodels>=0.14
python-dotenv>=1.0
requests>=2.31
